refactor remaining tests to use same consistent parameterised approach. Add logging to other remaining methods
