+++ testing-suite/staking-v4/caching.py
@@ -3,6 +3,8 @@ from config import DEFAULT_PROXY
 from chain_commander import add_blocks
 import time
 
+from utils.logger import logger
+
 
 def force_reset_validator_statistics():
     route = f"{DEFAULT_PROXY}/simulator/force-reset-validator-statistics"
@@ -14,3 +16,4 @@ def force_reset_validator_statistics():
 
     # wait 1 sec
     time.sleep(1)
+    logger.info("Validator statistics reset successfully and additional block added")
