+++ testing-suite/staking-v4/network_provider/key_management.py
@@ -4,8 +4,12 @@ from chain_commander import add_blocks_until_epoch_reached, add_blocks
 from config import proxy_default, DEFAULT_PROXY
 import json
 
+from utils.logger import logger
+
 
 def add_key(keys: list[ValidatorKey]) -> str:
+
+    logger.info("Adding keys to simulator")
     private_keys = []
     for key in keys:
         private_keys.append(key.get_private_key())
@@ -17,10 +21,12 @@ def add_key(keys: list[ValidatorKey]) -> str:
     json_structure = json.dumps(post_body)
     req = requests.post(f"{DEFAULT_PROXY}/simulator/add-keys", data=json_structure)
 
+    logger.info("Keys added successfully")
     return req.text
 
 
 def add_blocks_until_key_eligible(keys: list[ValidatorKey]) -> ValidatorKey:
+    logger.info("Attempting to reach eligibility for given keys")
     flag = False
     while not flag:
         for key in keys:
@@ -34,5 +40,6 @@ def add_blocks_until_key_eligible(keys: list[ValidatorKey]) -> ValidatorKey:
                 current_epoch = proxy_default.get_network_status().epoch_number
                 add_blocks_until_epoch_reached(current_epoch+1)
                 add_blocks(3)
-
+                
+    logger.info(f"Key {eligible_key.get_public_key()} is now eligible")
     return eligible_key
