+++ testing-suite/staking-v4/pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 testpaths = scenarios
 
-addopts = --tb=short
+addopts = --tb=short -s
